+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## Features
 
 -   Added a new `--sitemapBaseUrl` option. When specified, TypeDoc will generate a `sitemap.xml` in your output folder that describes the site, #2480.
+-   Added support for the `@class` tag. When added to a comment on a variable or function, TypeDoc will convert the member as a class, #2479.
+    Note: This should only be used on symbols which actually represent a class, but are not declared as a class for some reason.
 
 ## Bug Fixes
 
 -   Fixed an issue where a namespace would not be created for merged function-namespaces which are declared as variables, #2478.
+-   Variable functions which have construct signatures will no longer be converted as functions, ignoring the construct signatures.
 -   Fixed an issue where, if the index section was collapsed when loading the page, all content within it would be hidden until expanded, and a member visibility checkbox was changed.
 
 ## v0.25.7 (2024-01-08)